#162 메인 페이지 및 티켓팅 검색 구현
