+++ client/src/Components/Ticktes/ItemList.jsx
@@ -11,6 +11,7 @@ const ItemGrid = styled.div`
   display: grid;
   grid-template-columns: repeat(4, 1fr);
   grid-gap: 40px;
+  position: relative;
 
   @media screen and (max-width: ${breakpoint.mobile}) {
     grid-template-columns: repeat(3, 1fr);
@@ -23,14 +24,23 @@ const ItemGrid = styled.div`
     grid-gap: 50px;
     margin: 0 20px;
   }
+
+  .null_info {
+    position: absolute;
+    left: 40%;
+  }
 `;
 
 export default function ItemList({ data }) {
   return (
     <ItemGrid>
-      {data.map((item, index) => {
-        return <CardItem data={item} key={index} />;
-      })}
+      {data.length === 0 ? (
+        <p className="null_info">공연이 존재하지 않습니다</p>
+      ) : (
+        data.map((item, index) => {
+          return <CardItem data={item} key={index} />;
+        })
+      )}
     </ItemGrid>
   );
 }
